Refactoring ConsoleOutputter and FileOutputter into StreamOutputter.
