+++ src/hayai_benchmarker.hpp
@@ -11,9 +11,7 @@
 #include "hayai_test_factory.hpp"
 #include "hayai_test_descriptor.hpp"
 #include "hayai_test_result.hpp"
-#include "hayai_console_outputter.hpp"
-#include "hayai_file_outputter.hpp"
-
+#include "hayai_stream_outputter.hpp"
 
 namespace hayai
 {
